Tmp file flag

+++ onnxscript/ir/_external_data.py
@@ -223,6 +223,7 @@ def convert_tensors_to_external(
     path = os.path.join(base_path, relative_path)
     # Check if file path is valid, and create subsequent subdirectories within the path if they don't exist
     os.makedirs(os.path.dirname(path), exist_ok=True)
+    tmp_file_created = False
     # Check if file exists. Load pre-existing external data if it does.
     if os.path.exists(path):
         # Check if any tensor in the model is using the destination file
@@ -241,6 +242,7 @@ def convert_tensors_to_external(
                 os.makedirs(tmp_path, exist_ok=True)
                 # If exisiting external tensors are not loaded to memory, copy the external data to a temporary location
                 os.rename(path, os.path.join(tmp_path, relative_path))
+                tmp_file_created = True
                 for tensor in tensors:
                     if (
                         isinstance(tensor, _core.ExternalTensor)
@@ -273,7 +275,7 @@ def convert_tensors_to_external(
 
     # Clean-up temporary file if it is created
     tmp_path = os.path.join(base_path, "tmp", relative_path)
-    if os.path.exists(tmp_path):
+    if os.path.exists(tmp_path) and tmp_file_created:
         os.remove(tmp_path)
 
     return external_tensors
